+++ internal/core/config_test.go
@@ -80,7 +80,10 @@ func createExpectedValidConfig(transferConfig transfertask.TransferConfig) Confi
 			},
 		},
 		PathsConf: wsconfig.PathsConf{
-			CollectionLocation: "/some/path",
+			CollectionLocations: map[string]string{
+				"path":   "/some/path",
+				"folder": "/another/folder",
+			},
 		},
 		MetadataExtJobsConf: wsconfig.MetadataExtJobsConf{
 			ConcurrencyLimit: 100,